Python: Removed locals() argument in eval/exec
This fixes assigned variables being out of scope in list comprehensions.
For example: `fn.py 'a = 1' '[a for _ in range(5)]'` will now correctly
return a five-element list of `1`s. This was happening because `a` was
being assigned to the local scope of the execution. A workaround was
labelling `a` as `global`: `global a; a = 1`. This workaround is not
needed anymore.